requirements.txt
pygame-ce
zengl
